# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PayIn Platform",
        "description": "API del componente PayIn: procesamiento de ingresos de fondos.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/api",
            "description": "Servidor de desarrollo"
        }
    ],
    "paths": {
...
        "/v1/payins": {
            "get": {
                "tags": [
                    "PayIns"
                ],
                "summary": "Listar PayIns",
                "description": "Lista PayIns con filtros y paginación.",
                "operationId": "6a63a358d46ccf0dbc5578dfaacf4442",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "created",
                                "validated",
                                "processing",
                                "processed",
                                "failed"
                            ]
                        }
                    },
                    {
                        "name": "from",
                        "in": "query",
                        "description": "Fecha inicial (ISO 8601)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date-time"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "description": "Fecha final (ISO 8601)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date-time"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 20,
                            "maximum": 100
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0
                        }
                    },
                    {
                        "name": "X-Correlation-Id",
                        "in": "header",
                        "description": "Identificador de correlación para trazabilidad",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Página de PayIns",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PayInPage"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "PayIns"
                ],
                "summary": "Procesar un PayIn",
                "description": "Crea y procesa un PayIn.\n\nLa operación valida los aggregates, persiste el PayIn (CREATED →\nVALIDATED), delega el cobro al proveedor resuelto por el método de\npago y aplica el resultado (PROCESSED/FAILED). Si el proveedor\nrechaza la operación, el PayIn se crea con estado FAILED y el\nerror_code correspondiente.",
                "operationId": "bf4900ce65a5a83fb780a2587ef7e610",
                "parameters": [
                    {
                        "name": "X-Correlation-Id",
                        "in": "header",
                        "description": "Identificador de correlación para trazabilidad",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StorePayInRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "PayIn creado y procesado (estado processed o failed según el proveedor)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PayInResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Datos de entrada inválidos o invariantes de dominio violadas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Cliente, cuenta, método de pago o proveedor inexistente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Referencia ya utilizada o conflicto de estado/concurrencia",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "502": {
                        "description": "Error inesperado del proveedor de pago",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/payins/{id}": {
            "get": {
                "tags": [
                    "PayIns"
                ],
                "summary": "Consultar un PayIn",
                "description": "Consulta un PayIn por su identificador.",
                "operationId": "f757657cd5b9f5010225c51e608835a7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID del PayIn",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "019fd738-d4f2-7bc9-916b-0ae687b42038"
                        }
                    },
                    {
                        "name": "X-Correlation-Id",
                        "in": "header",
                        "description": "Identificador de correlación para trazabilidad",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PayIn encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PayInResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "PayIn inexistente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ErrorResponse": {
                "required": [
                    "errors"
                ],
                "properties": {
                    "errors": {
                        "description": "Schema OpenAPI del envelope de error homogéneo de la API.",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Error"
                        }
                    }
                },
                "type": "object"
            },
            "Error": {
                "required": [
                    "code",
                    "message"
                ],
                "properties": {
                    "code": {
                        "description": "Código de error estable",
                        "type": "string",
                        "example": "REFERENCE_ALREADY_USED"
                    },
                    "message": {
                        "description": "Mensaje legible del error",
                        "type": "string",
                        "example": "La referencia \"postman-0001\" ya fue utilizada en otra operación."
                    },
                    "meta": {
                        "description": "Contexto adicional del error",
                        "type": "object",
                        "additionalProperties": true
                    }
                },
                "type": "object"
            },
...
            "PayInPage": {
                "required": [
                    "data",
                    "meta"
                ],
                "properties": {
                    "data": {
                        "description": "Schema OpenAPI de la página de listado de PayIns.",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PayIn"
                        }
                    },
                    "meta": {
                        "required": [
                            "total",
                            "limit",
                            "offset"
                        ],
                        "properties": {
                            "total": {
                                "type": "integer",
                                "example": 42
                            },
                            "limit": {
                                "type": "integer",
                                "example": 20
                            },
                            "offset": {
                                "type": "integer",
                                "example": 0
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "PayInResponse": {
                "required": [
                    "data"
                ],
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/PayIn"
                    }
                },
                "type": "object"
            },
            "PayIn": {
                "required": [
                    "id",
                    "client_id",
                    "account_id",
                    "payment_method_id",
                    "amount",
                    "currency",
                    "status",
                    "created_at"
                ],
                "properties": {
                    "id": {
                        "description": "Schema OpenAPI de la representación de un PayIn.",
                        "type": "string",
                        "format": "uuid",
                        "example": "019fd738-d4f2-7bc9-916b-0ae687b42038"
                    },
                    "client_id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "019fd715-ebf8-7223-ada8-b3c168a28e22"
                    },
                    "account_id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "019fd715-ec1a-7a7e-ab6f-f497aa52abe4"
                    },
                    "payment_method_id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "019fd715-ec43-784b-97dd-9b2fe70bfe69"
                    },
                    "amount": {
                        "type": "integer",
                        "example": 25000
                    },
                    "currency": {
                        "type": "string",
                        "enum": [
                            "COP",
                            "USD",
                            "EUR",
                            "MXN"
                        ],
                        "example": "COP"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "created",
                            "validated",
                            "processing",
                            "processed",
                            "failed"
                        ],
                        "example": "processed"
                    },
                    "reference": {
                        "type": "string",
                        "example": "order-2026-0001",
                        "nullable": true
                    },
                    "provider_id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "019fd715-eb24-7683-b8d2-9d83ffdca22d",
                        "nullable": true
                    },
                    "provider_transaction_id": {
                        "type": "string",
                        "example": "FP-019FD738-D4F",
                        "nullable": true
                    },
                    "error_code": {
                        "type": "string",
                        "example": null,
                        "nullable": true
                    },
                    "error_message": {
                        "type": "string",
                        "example": null,
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-08-06T12:54:23Z"
                    },
                    "processed_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-08-06T12:54:23Z",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "StorePayInRequest": {
                "required": [
                    "client_id",
                    "account_id",
                    "payment_method_id",
                    "amount",
                    "currency"
                ],
                "properties": {
                    "client_id": {
                        "description": "UUID del cliente que origina la operación",
                        "type": "string",
                        "format": "uuid",
                        "example": "019fd715-ebf8-7223-ada8-b3c168a28e22"
                    },
                    "account_id": {
                        "description": "UUID de la cuenta destino",
                        "type": "string",
                        "format": "uuid",
                        "example": "019fd715-ec1a-7a7e-ab6f-f497aa52abe4"
                    },
                    "payment_method_id": {
                        "description": "UUID del método de pago",
                        "type": "string",
                        "format": "uuid",
                        "example": "019fd715-ec43-784b-97dd-9b2fe70bfe69"
                    },
                    "amount": {
                        "description": "Monto en unidades menores (enteras) de la moneda",
                        "type": "integer",
                        "minimum": 1,
                        "example": 25000
                    },
                    "currency": {
                        "description": "Moneda ISO 4217",
                        "type": "string",
                        "enum": [
                            "COP",
                            "USD",
                            "EUR",
                            "MXN"
                        ],
                        "example": "COP"
                    },
                    "reference": {
                        "description": "Referencia idempotente del cliente (opcional)",
                        "type": "string",
                        "example": "order-2026-0001"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}